Clean up the quantile patching method
