+++ timeseries/src/autogluon/timeseries/models/chronos/pipeline/__init__.py
@@ -1,10 +1,11 @@
 from .base import BaseChronosPipeline, ForecastType
 from .chronos import ChronosPipeline
-from .chronos_bolt import ChronosBoltPipeline
+from .chronos_bolt import ChronosBoltPipeline, patch_chronos_bolt_output_quantiles
 
 __all__ = [
     "BaseChronosPipeline",
     "ChronosBoltPipeline",
     "ChronosPipeline",
     "ForecastType",
+    "patch_chronos_bolt_output_quantiles",
 ]
